Ignition v2 spec v3
change dir where ssh keys are written

first pass brute force substitution 2.2-> 3.0

vendor bump

first pass updating code for ign3 changes

ign v2->v3: update tests

update golden files

fix vsphere typo

Revert "update golden files"

This reverts commit 7a0beb3dc4cec04022315b68ec6093f70cd53f0c.

fix units

+++ pkg/daemon/update.go
@@ -15,12 +15,13 @@ import (
 	"syscall"
 	"time"
 
-	ignv2_2types "github.com/coreos/ignition/config/v2_2/types"
+	igntypes "github.com/coreos/ignition/config/v3_0/types"
 	"github.com/coreos/ignition/config/validate"
 	"github.com/golang/glog"
 	"github.com/google/renameio"
 	drain "github.com/openshift/kubernetes-drain"
 	mcfgv1 "github.com/openshift/machine-config-operator/pkg/apis/machineconfiguration.openshift.io/v1"
+	ctrlcommon "github.com/openshift/machine-config-operator/pkg/controller/common"
 	"github.com/openshift/machine-config-operator/pkg/daemon/constants"
 	errors "github.com/pkg/errors"
 	"github.com/vincent-petithory/dataurl"
@@ -277,14 +278,6 @@ func (dn *Daemon) reconcilable(oldConfig, newConfig *mcfgv1.MachineConfig) error
 	// resources, and the mcc should've fully rendered those out before the
 	// config gets here.
 
-	// Networkd section
-
-	// we don't currently configure the network in place. we can't fix it if
-	// something changed here.
-	if !reflect.DeepEqual(oldIgn.Networkd, newIgn.Networkd) {
-		return errors.New("ignition networkd section contains changes")
-	}
-
 	// Passwd section
 
 	// we don't currently configure Groups in place. we don't configure Users except
@@ -341,7 +334,7 @@ func (dn *Daemon) reconcilable(oldConfig, newConfig *mcfgv1.MachineConfig) error
 	// Special case files append: if the new config wants us to append, then we
 	// have to force a reprovision since it's not idempotent
 	for _, f := range newIgn.Storage.Files {
-		if f.Append {
+		if len(f.Append) != 0 {
 			return fmt.Errorf("ignition file %v includes append", f.Path)
 		}
 	}
@@ -360,8 +353,8 @@ func (dn *Daemon) reconcilable(oldConfig, newConfig *mcfgv1.MachineConfig) error
 // Otherwise, an error will be returned and the proposed config will not be reconcilable.
 // At this time we do not support non-"core" users or any changes to the "core" user
 // outside of SSHAuthorizedKeys.
-func verifyUserFields(pwdUser ignv2_2types.PasswdUser) error {
-	emptyUser := ignv2_2types.PasswdUser{}
+func verifyUserFields(pwdUser igntypes.PasswdUser) error {
+	emptyUser := igntypes.PasswdUser{}
 	tempUser := pwdUser
 	if tempUser.Name == coreUserName && len(tempUser.SSHAuthorizedKeys) >= 1 {
 		tempUser.Name = ""
@@ -483,7 +476,7 @@ func (dn *Daemon) deleteStaleData(oldConfig, newConfig *mcfgv1.MachineConfig) er
 }
 
 // enableUnit enables a systemd unit via symlink
-func (dn *Daemon) enableUnit(unit ignv2_2types.Unit) error {
+func (dn *Daemon) enableUnit(unit igntypes.Unit) error {
 	// The link location
 	wantsPath := filepath.Join(wantsPathSystemd, unit.Name)
 	// sanity check that we don't return an error when the link already exists
@@ -504,7 +497,7 @@ func (dn *Daemon) enableUnit(unit ignv2_2types.Unit) error {
 }
 
 // disableUnit disables a systemd unit via symlink removal
-func (dn *Daemon) disableUnit(unit ignv2_2types.Unit) error {
+func (dn *Daemon) disableUnit(unit igntypes.Unit) error {
 	// The link location
 	wantsPath := filepath.Join(wantsPathSystemd, unit.Name)
 	// sanity check so we don't return an error when the unit was already disabled
@@ -518,20 +511,20 @@ func (dn *Daemon) disableUnit(unit ignv2_2types.Unit) error {
 }
 
 // writeUnits writes the systemd units to disk
-func (dn *Daemon) writeUnits(units []ignv2_2types.Unit) error {
+func (dn *Daemon) writeUnits(units []igntypes.Unit) error {
 	for _, u := range units {
 		// write the dropin to disk
 		for i := range u.Dropins {
 			glog.Infof("Writing systemd unit dropin %q", u.Dropins[i].Name)
 			dpath := filepath.Join(pathSystemd, u.Name+".d", u.Dropins[i].Name)
-			if err := writeFileAtomicallyWithDefaults(dpath, []byte(u.Dropins[i].Contents)); err != nil {
+			if err := writeFileAtomicallyWithDefaults(dpath, []byte(ctrlcommon.StrFromStrPtr(u.Dropins[i].Contents))); err != nil {
 				return fmt.Errorf("failed to write systemd unit dropin %q: %v", u.Dropins[i].Name, err)
 			}
 
 			glog.V(2).Infof("Wrote systemd unit dropin at %s", dpath)
 		}
 
-		if u.Contents == "" {
+		if ctrlcommon.StrFromStrPtr(u.Contents) == "" {
 			continue
 		}
 
@@ -541,7 +534,7 @@ func (dn *Daemon) writeUnits(units []ignv2_2types.Unit) error {
 
 		// check if the unit is masked. if it is, we write a symlink to
 		// /dev/null and continue
-		if u.Mask {
+		if u.Mask != nil && *u.Mask {
 			glog.V(2).Info("Systemd unit masked")
 			if err := os.RemoveAll(fpath); err != nil {
 				return fmt.Errorf("failed to remove unit %q: %v", u.Name, err)
@@ -557,7 +550,7 @@ func (dn *Daemon) writeUnits(units []ignv2_2types.Unit) error {
 		}
 
 		// write the unit to disk
-		if err := writeFileAtomicallyWithDefaults(fpath, []byte(u.Contents)); err != nil {
+		if err := writeFileAtomicallyWithDefaults(fpath, []byte(ctrlcommon.StrFromStrPtr(u.Contents))); err != nil {
 			return fmt.Errorf("failed to write systemd unit %q: %v", u.Name, err)
 		}
 
@@ -571,12 +564,6 @@ func (dn *Daemon) writeUnits(units []ignv2_2types.Unit) error {
 		// be fine as disableUnit is idempotent.
 		// Note: we have to check for legacy unit.Enable and honor it
 		glog.Infof("Enabling systemd unit %q", u.Name)
-		if u.Enable {
-			if err := dn.enableUnit(u); err != nil {
-				return err
-			}
-			glog.V(2).Infof("Enabled systemd unit %q", u.Name)
-		}
 		if u.Enabled != nil {
 			if *u.Enabled {
 				if err := dn.enableUnit(u); err != nil {
@@ -596,11 +583,10 @@ func (dn *Daemon) writeUnits(units []ignv2_2types.Unit) error {
 
 // writeFiles writes the given files to disk.
 // it doesn't fetch remote files and expects a flattened config file.
-func (dn *Daemon) writeFiles(files []ignv2_2types.File) error {
+func (dn *Daemon) writeFiles(files []igntypes.File) error {
 	for _, file := range files {
 		glog.Infof("Writing file %q", file.Path)
-
-		contents, err := dataurl.DecodeString(file.Contents.Source)
+		contents, err := dataurl.DecodeString(ctrlcommon.StrFromStrPtr(file.Contents.Source))
 		if err != nil {
 			return err
 		}
@@ -612,7 +598,7 @@ func (dn *Daemon) writeFiles(files []ignv2_2types.File) error {
 			uid, gid = -1, -1
 		)
 		// set chown if file information is provided
-		if file.User != nil || file.Group != nil {
+		if (file.User != igntypes.NodeUser{}) || (file.Group != igntypes.NodeGroup{}) {
 			uid, gid, err = getFileOwnership(file)
 			if err != nil {
 				return fmt.Errorf("failed to retrieve file ownership for file %q: %v", file.Path, err)
@@ -626,37 +612,37 @@ func (dn *Daemon) writeFiles(files []ignv2_2types.File) error {
 }
 
 // This is essentially ResolveNodeUidAndGid() from Ignition; XXX should dedupe
-func getFileOwnership(file ignv2_2types.File) (int, int, error) {
+func getFileOwnership(file igntypes.File) (int, int, error) {
 	uid, gid := 0, 0 // default to root
-	if file.User != nil {
+	if (file.User != igntypes.NodeUser{}) {
 		if file.User.ID != nil {
 			uid = *file.User.ID
-		} else if file.User.Name != "" {
-			osUser, err := user.Lookup(file.User.Name)
+		} else if ctrlcommon.StrFromStrPtr(file.User.Name) != "" {
+			osUser, err := user.Lookup(ctrlcommon.StrFromStrPtr(file.User.Name))
 			if err != nil {
-				return uid, gid, fmt.Errorf("failed to retrieve UserID for username: %s", file.User.Name)
+				return uid, gid, fmt.Errorf("failed to retrieve UserID for username: %s", *file.User.Name)
 			}
-			glog.V(2).Infof("Retrieved UserId: %s for username: %s", osUser.Uid, file.User.Name)
+			glog.V(2).Infof("Retrieved UserId: %s for username: %s", osUser.Uid, *file.User.Name)
 			uid, _ = strconv.Atoi(osUser.Uid)
 		}
 	}
-	if file.Group != nil {
+	if (file.Group != igntypes.NodeGroup{}) {
 		if file.Group.ID != nil {
 			gid = *file.Group.ID
-		} else if file.Group.Name != "" {
-			osGroup, err := user.LookupGroup(file.Group.Name)
+		} else if ctrlcommon.StrFromStrPtr(file.Group.Name) != "" {
+			osGroup, err := user.LookupGroup(ctrlcommon.StrFromStrPtr(file.Group.Name))
 			if err != nil {
-				return uid, gid, fmt.Errorf("failed to retrieve GroupID for group: %s", file.Group.Name)
+				return uid, gid, fmt.Errorf("failed to retrieve GroupID for group: %s", *file.Group.Name)
 			}
-			glog.V(2).Infof("Retrieved GroupID: %s for group: %s", osGroup.Gid, file.Group.Name)
+			glog.V(2).Infof("Retrieved GroupID: %s for group: %s", osGroup.Gid, *file.Group.Name)
 			gid, _ = strconv.Atoi(osGroup.Gid)
 		}
 	}
 	return uid, gid, nil
 }
 
-func (dn *Daemon) atomicallyWriteSSHKey(newUser ignv2_2types.PasswdUser, keys string) error {
-	authKeyPath := filepath.Join(coreUserSSHPath, "authorized_keys")
+func (dn *Daemon) atomicallyWriteSSHKey(newUser igntypes.PasswdUser, keys string) error {
+	authKeyPath := filepath.Join(coreUserSSHPath, "authorized_keys.d", "ignition")
 
 	// Keys should only be written to "/home/core/.ssh"
 	// Once Users are supported fully this should be writing to PasswdUser.HomeDir
@@ -672,7 +658,7 @@ func (dn *Daemon) atomicallyWriteSSHKey(newUser ignv2_2types.PasswdUser, keys st
 }
 
 // Update a given PasswdUser's SSHKey
-func (dn *Daemon) updateSSHKeys(newUsers []ignv2_2types.PasswdUser) error {
+func (dn *Daemon) updateSSHKeys(newUsers []igntypes.PasswdUser) error {
 	if len(newUsers) == 0 {
 		return nil
 	}
